# Speed up tag's to_field_bytes