Fixes #307.

```
go_base_10_16_bls12     time:   [3.1047 ms 3.1121 ms 3.1195 ms]
                        change: [-19.706% -19.469% -19.215%] (p = 0.00 < 0.05)
                        Performance has improved.

go_base_10_160_bls12    time:   [29.734 ms 29.819 ms 29.906 ms]
                        change: [-19.368% -19.070% -18.767%] (p = 0.00 < 0.05)
                        Performance has improved.

go_base_10_16_pasta_pallas
                        time:   [3.1728 ms 3.1768 ms 3.1808 ms]
                        change: [-18.493% -18.204% -17.952%] (p = 0.00 < 0.05)
                        Performance has improved.

go_base_10_160_pasta_pallas
                        time:   [30.472 ms 30.538 ms 30.610 ms]
                        change: [-18.865% -18.513% -18.159%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
  ```