@huitseeker 2023-03-24
@porcuquine Yup, I’m a dunce.